requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.27
chromadb==0.4.22
sentence-transformers==2.5.1
transformers==4.37.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.1.2+cpu
python-dotenv==1.0.1 